【Fix PIR Unittest No.84 BUAA】Deprecate test_standalone_executor_aot_choose_kernel (#66442)
* Update test_standalone_executor_aot_choose_kernel.py

* Rename test_standalone_executor_aot_choose_kernel.py to test_standalone_executor_aot_choose_kernel_deprecated.py

* Update CMakeLists.txt

* Update codestyle

* Update CMakeLists.txt

* Add files via upload

* Delete test/standalone_executor/test_standalone_executor_aot_choose_kernel_deprecated.py

* Update CMakeLists.txt
